refactor(pandas): remove pre_execute
BREAKING CHANGE: ibis.pandas.udf removed in favor of ibis.udf.vectorized
