+++ ibis/backends/pandas/dispatch.py
@@ -1,11 +1,7 @@
-from functools import partial
-
 from multipledispatch import Dispatcher
 
 import ibis.common.exceptions as com
 import ibis.expr.operations as ops
-from ibis.backends.base import BaseBackend
-from ibis.expr.scope import Scope
 
 from .trace import TraceTwoLevelDispatcher
 
@@ -28,36 +24,6 @@ def execute_node_without_scope(node, **kwargs):
             'source.'
         ).format(type(node).__name__)
     )
-
-
-pre_execute = Dispatcher(
-    'pre_execute',
-    doc="""\
-Given a node, compute a (possibly partial) scope prior to standard execution.
-
-Notes
------
-This function is useful if parts of the tree structure need to be executed at
-the same time or if there are other reasons to need to interrupt the regular
-depth-first traversal of the tree.
-""",
-)
-
-
-# Default returns an empty scope
-@pre_execute.register(ops.Node)
-@pre_execute.register(ops.Node, BaseBackend)
-def pre_execute_default(node, *clients, **kwargs):
-    return Scope()
-
-
-# Merge the results of all client pre-execution with scope
-@pre_execute.register(ops.Node, [BaseBackend])
-def pre_execute_multiple_clients(node, *clients, scope=None, **kwargs):
-    scope = scope.merge_scopes(
-        list(map(partial(pre_execute, node, scope=scope, **kwargs), clients))
-    )
-    return scope
 
 
 execute_literal = Dispatcher(
