+++ ibis/backends/pandas/execution/__init__.py
@@ -8,4 +8,5 @@ from .strings import *  # noqa: F401,F403
 from .structs import *  # noqa: F401,F403
 from .temporal import *  # noqa: F401,F403
 from .timecontext import *  # noqa: F401,F403
+from .udf import *  # noqa: F401,F403
 from .window import *  # noqa: F401,F403
