+++ ibis/backends/pandas/execution/udf.py
@@ -1,0 +1,102 @@
+"""APIs for creating user-defined element-wise, reduction and analytic
+functions.
+"""
+
+import pandas as pd
+from pandas.core.groupby import SeriesGroupBy
+
+import ibis.expr.operations as ops
+
+from ..aggcontext import Transform
+from ..dispatch import execute_node
+
+
+@execute_node.register(
+    ops.ElementWiseVectorizedUDF, SeriesGroupBy, [SeriesGroupBy]
+)
+def execute_udf_node_groupby(op, *args, **kwargs):
+    func = op.func
+
+    groupers = [
+        grouper
+        for grouper in (getattr(arg, 'grouper', None) for arg in args)
+        if grouper is not None
+    ]
+
+    # all grouping keys must be identical
+    assert all(groupers[0] == grouper for grouper in groupers[1:])
+
+    # we're performing a scalar operation on grouped column, so
+    # perform the operation directly on the underlying Series
+    # and regroup after it's finished
+    args = [getattr(arg, 'obj', arg) for arg in args]
+    groupings = groupers[0].groupings
+    return func(*args).groupby(groupings)
+
+
+# Define an execution rule for a simple elementwise Series
+# function
+@execute_node.register(ops.ElementWiseVectorizedUDF, [pd.Series])
+def execute_udf_node(op, *args, **kwargs):
+    # We have rewritten op.func to be a closure enclosing
+    # the kwargs, and therefore, we do not need to pass
+    # kwargs here. This is true for all udf execution in this
+    # file.
+    # See ibis.udf.vectorized.UserDefinedFunction
+    return op.func(*args)
+
+
+# An execution rule to handle analytic and reduction UDFs over
+# 1) an ungrouped window,
+# 2) an ungrouped Aggregate node, or
+# 3) an ungrouped custom aggregation context
+@execute_node.register(
+    (ops.AnalyticVectorizedUDF, ops.ReductionVectorizedUDF),
+    pd.Series,
+    [pd.Series],
+)
+def execute_udaf_node_no_groupby(op, *args, aggcontext, **kwargs):
+    func = op.func
+    return aggcontext.agg(args[0], func, *args[1:])
+
+
+# An execution rule to handle analytic and reduction UDFs over
+# 1) a grouped window,
+# 2) a grouped Aggregate node, or
+# 3) a grouped custom aggregation context
+@execute_node.register(
+    (ops.AnalyticVectorizedUDF, ops.ReductionVectorizedUDF),
+    SeriesGroupBy,
+    [SeriesGroupBy],
+)
+def execute_udaf_node_groupby(op, *args, aggcontext, **kwargs):
+    func = op.func
+    if isinstance(aggcontext, Transform):
+        # We are aggregating over an unbounded (and GROUPED) window,
+        # which uses a Transform aggregation context.
+        # We need to do some pre-processing to func and args so that
+        # Transform can pull data out of the SeriesGroupBys in args.
+
+        # Construct a generator that yields the next group of data
+        # for every argument excluding the first (pandas performs
+        # the iteration for the first argument) for each argument
+        # that is a SeriesGroupBy.
+        iters = ((data for _, data in arg) for arg in args[1:])
+
+        # TODO: Unify calling convension here to be more like
+        # window
+        def aggregator(first, *rest):
+            # map(next, *rest) gets the inputs for the next group
+            # TODO: might be inefficient to do this on every call
+            return func(first, *map(next, rest))
+
+        return aggcontext.agg(args[0], aggregator, *iters)
+    else:
+        # We are either:
+        # 1) Aggregating over a bounded window, which uses a Window
+        #  aggregation context
+        # 2) Aggregating over a custom aggregation context
+        # 3) Aggregating using an Aggregate node (with GROUPING), which
+        #   uses a Summarize aggregation context
+        # No pre-processing to be done for any case.
+        return aggcontext.agg(args[0], func, *args[1:])
