+++ ibis/backends/pandas/tests/test_udf.py
@@ -8,8 +8,8 @@ import pytest
 import ibis
 import ibis.expr.datatypes as dt
 import ibis.expr.types as ir
+import ibis.udf.vectorized as udf
 from ibis.backends.pandas import Backend
-from ibis.backends.pandas.udf import udf
 
 
 @pytest.fixture
